+++ examples/poetry/src/koto_bindings.rs
@@ -2,10 +2,9 @@ use {
     crate::Poetry,
     koto::runtime::{
         make_runtime_error, runtime_error, ExternalData, ExternalIterator, ExternalValue, MetaMap,
-        RwLock, Value, ValueIterator, ValueIteratorOutput, ValueMap,
+        Value, ValueIterator, ValueIteratorOutput, ValueMap,
     },
-    lazy_static::lazy_static,
-    std::{fmt, sync::Arc},
+    std::{cell::RefCell, fmt, rc::Rc},
 };
 
 pub fn make_module() -> ValueMap {
@@ -29,41 +28,40 @@ pub fn make_module() -> ValueMap {
     result
 }
 
-lazy_static! {
-    static ref POETRY_BINDINGS: Arc<RwLock<MetaMap>> = {
-        use Value::{Empty, Iterator, Str};
+thread_local!(
+static POETRY_BINDINGS: Rc<RefCell<MetaMap>> = {
+    use Value::{Empty, Iterator, Str};
 
-        let mut bindings = MetaMap::with_type_name("Poetry");
+    let mut bindings = MetaMap::with_type_name("Poetry");
 
-        bindings.add_named_instance_fn_mut(
-            "add_source_material",
-            |poetry: &mut KotoPoetry, _, args| match args {
-                [Str(text)] => {
-                    poetry.0.add_source_material(text);
-                    Ok(Empty)
-                }
-                _ => runtime_error!("poetry.add_source_material: Expected a String as argument"),
-            },
-        );
+    bindings.add_named_instance_fn_mut(
+        "add_source_material",
+        |poetry: &mut KotoPoetry, _, args| match args {
+            [Str(text)] => {
+                poetry.0.add_source_material(text);
+                Ok(Empty)
+            }
+            _ => runtime_error!("poetry.add_source_material: Expected a String as argument"),
+        },
+    );
 
-        bindings.add_named_instance_fn("iter", |_poetry: &KotoPoetry, external_value, _| {
-            let iter = PoetryIter {
-                poetry: external_value.clone(),
-            };
-            Ok(Iterator(ValueIterator::make_external(iter)))
-        });
+    bindings.add_named_instance_fn("iter", |_poetry: &KotoPoetry, external_value, _| {
+        let iter = PoetryIter {
+            poetry: external_value.clone(),
+        };
+        Ok(Iterator(ValueIterator::make_external(iter)))
+    });
 
-        bindings.add_named_instance_fn_mut("next_word", |poetry: &mut KotoPoetry, _, _| {
-            let result = match poetry.0.next_word() {
-                Some(word) => Str(word.as_ref().into()),
-                None => Empty,
-            };
-            Ok(result)
-        });
+    bindings.add_named_instance_fn_mut("next_word", |poetry: &mut KotoPoetry, _, _| {
+        let result = match poetry.0.next_word() {
+            Some(word) => Str(word.as_ref().into()),
+            None => Empty,
+        };
+        Ok(result)
+    });
 
-        Arc::new(RwLock::new(bindings))
-    };
-}
+    Rc::new(RefCell::new(bindings))
+});
 
 #[derive(Clone)]
 struct PoetryIter {
@@ -82,7 +80,7 @@ impl Iterator for PoetryIter {
     fn next(&mut self) -> Option<Self::Item> {
         use Value::{Empty, Str};
 
-        match self.poetry.data.write().downcast_mut::<KotoPoetry>() {
+        match self.poetry.data.borrow_mut().downcast_mut::<KotoPoetry>() {
             Some(poetry) => {
                 let result = match poetry.0.next_word() {
                     Some(word) => Str(word.as_ref().into()),
@@ -102,8 +100,10 @@ pub struct KotoPoetry(Poetry);
 
 impl KotoPoetry {
     fn make_external_value(poetry: Poetry) -> Value {
-        let result =
-            ExternalValue::with_shared_meta_map(KotoPoetry(poetry), POETRY_BINDINGS.clone());
+        let result = ExternalValue::with_shared_meta_map(
+            KotoPoetry(poetry),
+            POETRY_BINDINGS.with(|meta| meta.clone()),
+        );
 
         Value::ExternalValue(result)
     }