+++ src/koto/tests/koto_tests.rs
@@ -117,13 +117,11 @@ assert_near 1, 2, 0.1
     koto_test!(number_ops);
     koto_test!(numbers);
     koto_test!(num2_4);
-    koto_test!(os);
     koto_test!(primes);
     koto_test!(ranges);
     koto_test!(strings);
     koto_test!(string_formatting);
     koto_test!(tests);
-    koto_test!(threads);
     koto_test!(tuples);
     koto_test!(types);
 }